+++ src/example/cifar/test_cifar.cc
@@ -20,6 +20,7 @@
 #include "paddle/fluid/platform/gpu_info.h"
 #include "paddle/fluid/platform/place.h"
 #include "src/function.h"
+#include "src/optimizer.h"
 
 using paddle::tape::VariableHandle;
 using paddle::tape::Linear;
@@ -34,6 +35,7 @@ using paddle::tape::cross_entropy;
 using paddle::tape::reset_global_tape;
 using paddle::tape::get_global_tape;
 using paddle::tape::OptimizableParameters;
+using paddle::tape::BackwardAndUpdate;
 
 using paddle::tape::CreateRecordioFileReader;
 using paddle::tape::ReadNext;
@@ -156,7 +158,7 @@ TEST(Cifar, TestGPU) {
     auto loss = mean(cross_entropy(predict, label));
     auto precision = accuracy(predict, label);
 
-    get_global_tape().BackwardAndUpdate(loss, &adam);
+    BackwardAndUpdate(loss, &adam);
 
     // Every time certain amount of batches have been processed,
     // we test the average loss and accuracy on the test data set,
